app/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
Django==5.2
django-extensions==4.1
djangorestframework==3.16.0
executing==2.2.0
Faker==37.1.0
gunicorn==23.0.0
idna==3.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
lxml==5.4.0
matplotlib-inline==0.1.7
packaging==25.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.7
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
